requirements.txt
pandas
scikit-learn
pip
tldextract
requests
python-whois 
beautifulsoup4